requirements.txt
"fastapi[all]"
openai
python-dotenv
google-generativeai
"sqlalchemy[pydantic-compat]"
